requirements.txt
anyio==3.6.2
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
bcrypt==4.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
decorator==5.1.1
ecdsa==0.18.0
executing==1.1.1
fastapi==0.85.1
greenlet==1.1.3.post0
h11==0.14.0
httptools==0.5.0
idna==3.4
iniconfig==1.1.1
jedi==0.18.1
matplotlib-inline==0.1.6
packaging==21.3
parso==0.8.3
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
pluggy==1.0.0
prompt-toolkit==3.0.31
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
Pygments==2.13.0
pyparsing==3.0.9
pytest==7.2.0
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.42
stack-data==0.5.1
starlette==0.20.4
traitlets==5.5.0
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
uvloop==0.17.0
watchfiles==0.18.0
wcwidth==0.2.5
websockets==10.4
yapf==0.32.0
